% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression}
\alias{expression}
\title{My Gene Expression table}
\format{
The 'expression' dataset is a tibble of gene expression counts of samples and gene
\itemize{
\item 'gene': HUGO gene symbol
\item 'sample': This is a biological
\item 'counts': absolute raw counts
}
}
\usage{
expression
}
\description{
My Gene Expression table
}
\keyword{datasets}
