% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_char.R
\name{factor_to_char}
\alias{factor_to_char}
\title{Length of the factor equals number of levels}
\usage{
factor_to_char(x)
}
\arguments{
\item{x}{factor}
}
\value{
logical
}
\description{
Length of the factor equals number of levels
}
\examples{
factor_to_char(iris$Species)
}

