% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors together}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Bind two factors together
}
\examples{
fbind(iris$Species[c(1,51,101)],PlantGrowth$group[c(1,11,21)])
}

