% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.R
\name{is_odd}
\alias{is_odd}
\title{Check whether a number is odd
@description
This function takes a numeric vector and performs a modulo to get the division
remainder and compares it to zero \eqn{\alpha}}
\usage{
is_odd(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Returns a logical vector that indicates whether the numbers in  or not
}
\description{
Check whether a number is odd
@description
This function takes a numeric vector and performs a modulo to get the division
remainder and compares it to zero \eqn{\alpha}
}
\examples{
is_odd(1)
is_odd(2)


}
