% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_shannon.R
\name{my_shannon}
\alias{my_shannon}
\title{Shannon Diversity Index}
\usage{
my_shannon(x)
}
\arguments{
\item{x}{A vector with values of abundance in the community}
}
\value{
A numeric vector of length 1 with the value of the Shannon diversity
}
\description{
Custom function to calculate shannon diversity index
}
\examples{
com <- c(1, 2, 3)
my_shannon(com)
}
