% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_levels.R
\name{reorder_levels}
\alias{reorder_levels}
\title{Reorder factor levels based on frequency}
\usage{
reorder_levels(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Reorder factor levels based on frequency
}
\examples{
reorder_levels(factor(c("a","a","b","c","c","c","c","d","d","d")))
}

