% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.R
\name{is_odd}
\alias{is_odd}
\title{Check whether a number is odd}
\usage{
is_odd(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A logical vector that calculate the modulo to zero if number x is odd or not
}
\description{
Check whether a number is odd
}
\examples{
is_odd(3)
is_odd(2)



}
